machine-lab-manager/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
aio-pika==9.4.0
passlib[bcrypt]==1.7.4
PyJWT[crypto]==2.8.0